Refactored css files into a director, and component files into a directory.
